+++ CHANGELOG.md
@@ -16,6 +16,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ---
 
+## [0.6.1] - 2023-09-22
+
+### Fixed
+* Implement actual search in AkoSearchSelectionSheet
+
 ## [0.6.0] - 2023-09-08
 
 ### Added